RD-server: don't set OTHER flag if the DHCP server is not running
The RD-server sets ND_RA_FLAG_OTHER even if the DHCP server is
not runnng, causing hosts on the link to fire a dhcp client and
try to get additional configuration parameters forever.

This is mostly harmless as SLAAC is working as expected (with DNS
addresses obtained through RAs), but generates unnecessary multicast
traffic on the link.

This commit fixes it by only setting the flag when the dhcp server
is running.